minor changes for better readability
